% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_profile.R
\name{create_profile}
\alias{create_profile}
\title{Create Profile Function}
\usage{
create_profile(name, overwrite = TRUE)
}
\arguments{
\item{name}{A string indicating the name of the profile to create.}

\item{overwrite}{A logical indicating whether to overwrite files if the
profile directory already exists.}
}
\description{
Create a new RStudio profile from the current global options using the
specified name.
}
