% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_profile.R
\name{createProfile}
\alias{createProfile}
\title{Create Profile Addin}
\usage{
createProfile()
}
\description{
Create a new RStudio profile from the current global options using the
user-specified name.
}
