% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_profile.R
\name{default_profile_addin}
\alias{default_profile_addin}
\title{Default Profile Addin}
\usage{
default_profile_addin()
}
\description{
Revert to the default profile configuration (with an RStudio prompt).
}
