% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_profiles.R
\name{list_profiles}
\alias{list_profiles}
\title{List Profiles Function}
\usage{
list_profiles(quiet = FALSE)
}
\arguments{
\item{quiet}{A flag indicating whether to print profiles to console.}
}
\value{
A character vector of RStudio profiles names.
}
\description{
List the current user's saved profiles.
}
