% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_profile.R
\name{load_profile}
\alias{load_profile}
\title{Load Profile Function}
\usage{
load_profile(name)
}
\arguments{
\item{name}{A string indicating the name of the profile to load.}
}
\description{
Apply the RStudio profile with the specified name.
}
