% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_profile.R
\name{open_profile}
\alias{open_profile}
\title{Open Profile Function}
\usage{
open_profile(name)
}
\arguments{
\item{name}{A string indicating the name of the profile to open.}
}
\description{
Open a saved profile and apply its configuration.
}
