% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles-package.R
\docType{package}
\name{profiles-package}
\alias{profiles}
\alias{profiles-package}
\title{profiles: Create and Switch Between RStudio Profiles}
\description{
Functions and RStudio addins for creating and quickly switching between RStudio "profiles," which contain user-specific preferences ranging from editor appearance to dictionaries and keybindings.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jmgirard/profiles}
  \item Report bugs at \url{https://github.com/jmgirard/profiles/issues}
}

}
\author{
\strong{Maintainer}: Jeffrey Girard \email{me@jmgirard.com} (\href{https://orcid.org/0000-0002-7359-3746}{ORCID})

}
\keyword{internal}
