% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_profile.R
\name{remove_profile}
\alias{remove_profile}
\title{Remove Profile Function}
\usage{
remove_profile(name, force = FALSE)
}
\arguments{
\item{name}{A string indicating the name of the profile to remove.}

\item{force}{A flag indicating whether permissions should be changed (if
possible) to allow the directory and its files to be removed.}
}
\description{
Remove a saved profile.
}
