% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_profile.R
\name{save_profile}
\alias{save_profile}
\title{Save Profile Function}
\usage{
save_profile(name, overwrite = TRUE)
}
\arguments{
\item{name}{A string indicating the name of the profile to create.}

\item{overwrite}{A logical indicating whether to overwrite files if the
profile directory already exists.}
}
\description{
Save the current configuration as a profile.
}
