% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_profiles.R
\name{list_profiles}
\alias{list_profiles}
\title{List Profiles Function}
\usage{
list_profiles()
}
\value{
A character vector of RStudio profiles names.
}
\description{
Find the names of all RStudio profiles and return as a character vector.
}
